// microshop-ui/src/app/api/order/model/create-order-dto.ts
/**
 * MicroShop.OrderAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CreditCardDto } from './credit-card-dto';
import { OrderItemDto } from './order-item-dto';
import { PaymentType } from './payment-type';
import { AddressDto } from './address-dto';


export interface CreateOrderDto { 
    userId?: string;
    address?: AddressDto;
    items?: Array<OrderItemDto> | null;
    couponCode?: string | null;
    paymentType?: PaymentType;
    cardInfo?: CreditCardDto;
}
export namespace CreateOrderDto {
}