// microshop-ui/src/app/api/order/model/address-dto.ts
/**
 * MicroShop.OrderAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AddressDto { 
    line?: string | null;
    city?: string | null;
    district?: string | null;
    zipCode?: string | null;
}