// microshop-ui/src/app/api/order/model/credit-card-dto.ts
/**
 * MicroShop.OrderAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreditCardDto { 
    cardNumber?: string | null;
    holderName?: string | null;
    expiration?: string | null;
    cvv?: string | null;
}