// microshop-ui/src/app/api/order/model/payment-type.ts
/**
 * MicroShop.OrderAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const PaymentType = {
    NUMBER_0: 0,
    NUMBER_1: 1
} as const;
export type PaymentType = typeof PaymentType[keyof typeof PaymentType];